fix: Consider all amounts in multi-currency postings
For a transaction like:

```ledger
2024/01/01 Cash transfer
    Assets:Savings
    Assets:Checking                          -20,000 INR
    Assets:Checking                             -500 USD
```

The posting returned by `hledger` will have a list of two amounts in
the JSON key `pamount`, for the `Assets:Savings` account. Currently, we
are only consider the first one we find, so the one for USD is ignored.

This change iterates over all the received amounts for each posting, to
avoid incorrect calculations when this type of transactions are present.